:bug: fix: Array de soluções para float
Alteração do tipo e das operações sob o array de soluções para float.

+++ Header.h
@@ -20,7 +20,7 @@ typedef struct Tableaux{
     char bases[LNS][BFF];
     
     float values[LNS][CLS];
-    int solutions[CLS];
+    float solutions[CLS];
     int nvars, nexps, vstrl;
 } * Tableaux;
